Make efficient use of Primitives on CreationManager
I've always known that this was an inefficient use of primitives
and that I was "drawing" more triangles than needed, but at the
time I originally wrote the piece of code, I couldn't figure out
why the card wasn't rendering as it should. Turned out the order
of the vertices would make the the two required triangles overlap
at the bottom and leave a gap at the top(although as the culling
was on I only say one triangle). My solution at the time was to
draw as many triangles until another triangle would appear and
fully draw the card.